fixed data mapping issue
